[OF-1831] build on linux (#959)
* [OF-1831] feat: Added platform macros and compiler pragmas to csp common

This required a new CSP_LINUX macro to identify the Linux platform + pragma warning macros for gcc. I also refactored this file by determining pragma warning macros by compiler instead of the platform

* [OF-1831] Added platform specific DeviceId implementation for Linux

This adds  Linux implementaiton for DeviceIdPath(). /.local/share/ is used for Linux as this is the user data is stored. I also refactored the include to seperate windows from POSIX headers

* [OF-1831] feat: Added gnu compile options

This mostly uses the clang compile options except for 2 instances: no-missing-field-initializers, is clang only, so this isn't used for gcc. no-error=changes-meaning was added to gcc the it was erroring due to types being the same name as members. This is a large change to remove that we have ticketed

* [OF-1831] feat: Fixing case sensitive errors for files/includes

gcc is stricter with case sensitivity. There were a few instances where out capitalizaiton was wrong

* [OF-1831] feat: Fixing incorrect namespacing issues

We were redeclaring the namespaces in certain places for declarations

* [OF-1831] feat: Adding missing includes

Fixing compile errors due to missing includes in gcc

* [OF-1831] feat: Fixing ambiguous constuctor issues

I had to explicity set int64_t over using ll postfix due to other similar constuctors

* [OF-1831] feat: Fixing incomplete type error with unique_ptr

By setting = nullptr, this was causing an initializaiton in the header which doesn't have the full type info for TinyProcessLib::Process

* [OF-1831] feat: Fixing not a string literal error

Failing due to prinf expecting a format string literal. This could have been repressed with format-security, but better to fix here

* [OF-1831] feat: Fixing trigraph error

We were using ??/ in our string which was creating a legacy trigraph. Prefixed the ? with a backlash to fix this

* [OF-1831] feat: Added linux gcc conan profile

* [OF-1831] feat: Fixing nullptr_t errors when building with clang

Building with clang caused nullptr_t to not be found due to different implementations. Now using the namespaced version to make it fully portable

* [OF-1831] feat: Adding linux clang profile

* [OF-1831] docs: Changelog

* [OF-1831] feat: Added missing sighandler typedef

sighandler_t is not part of standard C POSIX, so may not be defined, so just cerate the typedef ourselves

* [OF-1831] feat: Rocky Linux Fixes

Couple of changes to get csp building on Rocky Linux: 1. Update the clang path to not postfix the version 2. include exception in a signalr header for exception_ptr

* [OF-1831] docs: Added docs for dlopen divergence on Android

+++ Tests/src/PublicAPITests/ReplicatedValueTests.cpp
@@ -229,7 +229,7 @@ CSP_PUBLIC_TEST(CSPEngine, ReplicatedValueTests, MapAssignmentTest)
 // Tests move logic with a basic type
 CSP_PUBLIC_TEST(CSPEngine, ReplicatedValueTests, MoveConstructorIntTest)
 {
-    csp::common::ReplicatedValue Value { 10ll };
+    csp::common::ReplicatedValue Value { static_cast<int64_t>(10ll) };
     csp::common::ReplicatedValue NewValue { std::move(Value) };
 
     EXPECT_EQ(NewValue.GetReplicatedValueType(), csp::common::ReplicatedValueType::Integer);
@@ -252,8 +252,8 @@ CSP_PUBLIC_TEST(CSPEngine, ReplicatedValueTests, MoveConstructorStringTest)
 
 CSP_PUBLIC_TEST(CSPEngine, ReplicatedValueTests, MoveAssignmentIntTest)
 {
-    csp::common::ReplicatedValue Value { 10ll };
-    csp::common::ReplicatedValue NewValue { 5ll };
+    csp::common::ReplicatedValue Value { static_cast<int64_t>(10ll) };
+    csp::common::ReplicatedValue NewValue { static_cast<int64_t>(5ll) };
 
     NewValue = std::move(Value);
